:recycle: [REFACTOR] testTime Long -> Integer
testTime Long -> Integer

+++ src/main/java/swm_nm/morandi/test/domain/Test.java
@@ -19,7 +19,7 @@ public class Test {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long testId;
     private LocalDateTime testDate;
-    private Long testTime;
+    private Integer testTime;
     private Integer problemCount;
 
     @Enumerated(EnumType.STRING)
