+++ src/main/java/swm_nm/morandi/test/dto/TestDto.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class TestDto {
     private Long testId;
     private LocalDateTime testDate;
-    private Long testTime;
+    private Integer testTime;
     private Integer problemCount;
     private DifficultyLevel startDifficulty;
     private DifficultyLevel endDifficulty;
